+++ stroom-query-jooq/src/main/java/stroom/query/jooq/QueryServiceJooqImpl.java
@@ -46,7 +46,10 @@ import java.util.Optional;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
-import static org.jooq.impl.DSL.*;
+import static org.jooq.impl.DSL.and;
+import static org.jooq.impl.DSL.field;
+import static org.jooq.impl.DSL.not;
+import static org.jooq.impl.DSL.or;
 
 public class QueryServiceJooqImpl<
         DOC_REF_ENTITY extends DocRefJooqEntity,
@@ -64,12 +67,13 @@ public class QueryServiceJooqImpl<
     private final Table<Record> table;
 
     @Inject
-    public QueryServiceJooqImpl(final QueryableEntity.ClassProvider<QUERYABLE_ENTITY> dtoClassProvider,
-                                final DocRefService<DOC_REF_ENTITY> docRefService,
-                                final Configuration jooqConfig) {
-        this.docRefService = docRefService;
+    @SuppressWarnings("unchecked")
+    public QueryServiceJooqImpl(final QueryableEntity.ClassProvider dtoClassProvider,
+                                final DocRefService docRefService,
+                                final DSLContext database) {
+        this.docRefService = (DocRefService<DOC_REF_ENTITY>) docRefService;
         this.dtoClass = dtoClassProvider.get();
-        this.database = DSL.using(jooqConfig);
+        this.database = database;
 
         this.table = Optional.ofNullable(dtoClass.getAnnotation(JooqEntity.class))
                 .map(JooqEntity::tableName)