+++ stroom-query-testing/src/integration-test/java/stroom/query/testing/jooq/app/JooqApp.java
@@ -1,6 +1,8 @@
 package stroom.query.testing.jooq.app;
 
 import com.codahale.metrics.health.HealthCheck;
+import com.google.inject.Guice;
+import com.google.inject.Injector;
 import io.dropwizard.Application;
 import io.dropwizard.configuration.EnvironmentVariableSubstitutor;
 import io.dropwizard.configuration.SubstitutingSourceProvider;
@@ -9,14 +11,12 @@ import io.dropwizard.setup.Environment;
 import stroom.query.jooq.AuditedJooqQueryBundle;
 
 public class JooqApp extends Application<JooqConfig> {
-    private final AuditedJooqQueryBundle<JooqConfig,
+    private Injector injector;
+
+    private AuditedJooqQueryBundle<JooqConfig,
             TestDocRefServiceJooqImpl,
             TestDocRefJooqEntity,
-            TestQueryableJooqEntity> auditedQueryBundle =
-            new AuditedJooqQueryBundle<>(
-                    TestDocRefServiceJooqImpl.class,
-                    TestDocRefJooqEntity.class,
-                    TestQueryableJooqEntity.class);
+            TestQueryableJooqEntity> auditedQueryBundle;
 
     public static void main(final String[] args) throws Exception {
         new JooqApp().run(args);
@@ -33,12 +33,22 @@ public class JooqApp extends Application<JooqConfig> {
             }
         });
 
-        environment.jersey().register(CreateTestDataJooqImpl.class);
+        environment.jersey().register(injector.getInstance(CreateTestDataJooqImpl.class));
     }
 
     @Override
     public void initialize(final Bootstrap<JooqConfig> bootstrap) {
         super.initialize(bootstrap);
+
+        auditedQueryBundle =
+                new AuditedJooqQueryBundle<>(
+                        (c) -> {
+                            injector = Guice.createInjector(auditedQueryBundle.getGuiceModule(c));
+                            return injector;
+                        },
+                        TestDocRefServiceJooqImpl.class,
+                        TestDocRefJooqEntity.class,
+                        TestQueryableJooqEntity.class);
 
         // This allows us to use templating in the YAML configuration.
         bootstrap.setConfigurationSourceProvider(new SubstitutingSourceProvider(