+++ stroom-query-testing/src/test/java/stroom/query/testing/TestFifoLogbackRule.java
@@ -8,7 +8,9 @@ import stroom.query.audit.logback.FifoLogbackAppender;
 
 import java.util.function.Predicate;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import static stroom.query.testing.FifoLogbackRule.containsAllOf;
 
 public class TestFifoLogbackRule {