+++ CHANGELOG.md
@@ -10,7 +10,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+## [v2.0.0-alpha.36] - 2018-03-26
+
+### Changed
+
+* jOOQ injecting DSL Context rather than configuration
+
+## [v2.0.0-alpha.35] - 2018-03-23
+
+### Changed
+
+* gh-13 - Using Guice for common auditable query resource implementations
+
 ## [v2.0.0-alpha.34] - 2018-03-23
+
+### Changed
 
 * Simplification of search completion state.
 
@@ -216,7 +230,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Initial release
 
-[Unreleased]: https://github.com/gchq/stroom-query/compare/v2.0.0-alpha.34...HEAD
+[Unreleased]: https://github.com/gchq/stroom-query/compare/v2.0.0-alpha.36...HEAD
+[v2.0.0-alpha.36]: https://github.com/gchq/stroom-query/compare/v2.0.0-alpha.35...v2.0.0-alpha.36
+[v2.0.0-alpha.35]: https://github.com/gchq/stroom-query/compare/v2.0.0-alpha.34...v2.0.0-alpha.35
 [v2.0.0-alpha.34]: https://github.com/gchq/stroom-query/compare/v2.0.0-alpha.33...v2.0.0-alpha.34
 [v2.0.0-alpha.33]: https://github.com/gchq/stroom-query/compare/v2.0.0-alpha.32...v2.0.0-alpha.33
 [v2.0.0-alpha.32]: https://github.com/gchq/stroom-query/compare/v2.0.0-alpha.31...v2.0.0-alpha.32